Remove problems UI (#94)
