+++ app/components/header/Feedback.tsx
@@ -1,7 +1,7 @@
 import { toast } from 'react-toastify';
 import ReactModal from 'react-modal';
 import { useState } from 'react';
-import { supabaseSubmitFeedback } from '~/lib/supabase/problems';
+import { supabaseSubmitFeedback } from '~/lib/supabase/feedback';
 import { getLastChatMessages } from '~/components/chat/Chat.client';
 
 ReactModal.setAppElement('#root');
